+++ tests/compute/test_unit.py
@@ -49,8 +49,8 @@ def test_planar_numpy():
 
 def test_spatial_object():
     v = vector.backends.object.VectorObject3D(
-        vector.backends.object.AzimuthalObjectXY(0.1, 0.2),
-        vector.backends.object.LongitudinalObjectZ(0.3),
+        azimuthal=vector.backends.object.AzimuthalObjectXY(0.1, 0.2),
+        longitudinal=vector.backends.object.LongitudinalObjectZ(0.3),
     )
     u = v.unit()
     assert type(u) is type(v)